requirements.txt
groq
python-dotenv
streamlit
flask


